Add mpsc::Sender argument to spawn_mount
Adds an Option<mpsc::Sender<()>> argument to spawn_mount and
spawn_mount2. If supplied, the background thread will send a unit
message on the channel when the thread terminates (e.g., by the user
unmounting the filesystem). This can be used in conjunction with
listening for a different user abort, such as Ctrl+C. See
https://github.com/cberner/fuser/issues/179

Note: the current implementation changes the function signatures of
spawn_mount and spawn_mount2.